refactor: VisualTransformation 파일 util 패키지로 이동
